requirements.txt
Django==1.8
Pillow==2.8.1
Unidecode==0.04.17
boto==2.38.0
django-appconf==1.0.1
django-bootstrap-form==3.2
django-compressor==1.5
django-debug-toolbar==1.3.0
django-storages==1.1.8
djangorestframework==3.1.1
drf-extensions==0.2.7
gnureadline==6.3.3
gunicorn==19.3.0
ipython==3.1.0
psycopg2==2.6
pycrypto==2.6.1
requests==2.6.2
six==1.9.0
sqlparse==0.1.15
wsgiref==0.1.2
